requirements.txt
streamlit
face_recognition
opencv-python-headless
Pillow
numpy
